+++ dyntapy/demand_data.py
@@ -215,7 +215,7 @@ def get_centroid_grid_coords(name: str, spacing=default_centroid_spacing):
     return x, y
 
 
-def add_centroids_to_graph(g, X, Y, k=1, method="turn", euclidean=False, **kwargs):
+def add_centroids(g, X, Y, k=1, method="turn", euclidean=False, **kwargs):
     """
     Adds centroids to g.
     Each centroids data dict contains 'x_coord','y_coord' and